+++ src/jvmMain/kotlin/tool/filter/GammaFilter.kt
@@ -5,7 +5,7 @@ import ru.nsu.ccfit.plum.draw.getImageFilter
 import kotlin.math.pow
 
 object GammaFilter : Filter("Гамма-коррекция") {
-    var gamma = 2.2
+    var gamma = 2.2f
 
     override fun permit(image: PlumImage): PlumImage {
         val newImage = image.copy()
@@ -14,9 +14,9 @@ object GammaFilter : Filter("Гамма-коррекция") {
             val pixel = newImage.getPixel(x, y)
 
             getNewColor(
-                com(pixel.first.toDouble().pow(1 / gamma).toInt()),
-                com(pixel.second.toDouble().pow(1 / gamma).toInt()),
-                com(pixel.third.toDouble().pow(1 / gamma).toInt())
+                com(pixel.first.toFloat().pow(1 / gamma).toInt()),
+                com(pixel.second.toFloat().pow(1 / gamma).toInt()),
+                com(pixel.third.toFloat().pow(1 / gamma).toInt())
             )
         }
